+++ src/OmegaBot.ts
@@ -1,4 +1,4 @@
-import { Client, TextChannel, Guild, GuildChannel, MessageOptions, Attachment, RichEmbed, WebhookMessageOptions, Permissions, Game } from "discord.js";
+import { Client, TextChannel, Guild, GuildChannel, MessageOptions, Attachment, RichEmbed, WebhookMessageOptions, Permissions, Game, Message } from "discord.js";
 import { cfg, LOGTAG } from "./config";
 import * as jdenticon from "jdenticon";
 
@@ -8,13 +8,24 @@ import { readFileSync, writeFileSync, unlinkSync, mkdirSync } from "fs";
 import { shuffle } from "./shuffle";
 import { stringify } from "querystring";
 
+interface GuildConfiguration {
+	allowAll: boolean,
+	streamerList: string[],
+	streamerChannelId: string,
+	announcementDelayHours: number,
+	announcerMessage: string,
+	botname: string
+}
+
 export class OmegaBot extends WorkerProcess {
 	protected DiscordBot: Client = null;
 	protected timer: NodeJS.Timer = null;
 	protected streamerChecks: Map<string, NodeJS.Timer> = new Map<string, NodeJS.Timer>();
-	protected streamerChannel: Map<string, string> = new Map<string, string>();
-	protected streamerAllowed: Map<string, string[]> = new Map<string, string[]>();
-	protected streamerAllowAll: Map<string, boolean> = new Map<string, boolean>();
+	protected guildConfigList: Map<string, GuildConfiguration> = new Map<string, GuildConfiguration>();
+
+	// protected streamerChannel: Map<string, string> = new Map<string, string>();
+	// protected streamerAllowed: Map<string, string[]> = new Map<string, string[]>();
+	// protected streamerAllowAll: Map<string, boolean> = new Map<string, boolean>();
 
 	protected announcementCache: Map<string, Map<string, Game>> = new Map<string, Map<string, Game>>();
 	protected announcementDateCache: Map<string, Map<string, Date>> = new Map<string, Map<string, Date>>();
@@ -42,23 +53,29 @@ export class OmegaBot extends WorkerProcess {
 	 * @param {string} guildId
 	 * @memberof OmegaBot
 	 */
-	protected saveGuildSettings(guildId: string) {
-		const GuildSettings = {
-			allowAll: this.streamerAllowAll.get(guildId) || false,
-			streamerList: this.streamerAllowed.get(guildId) || [],
-			streamerChannelId: this.streamerChannel.get(guildId) || null
+	protected saveGuildSettings(guildId: string, msg?: Message) {
+		const GuildConfig: GuildConfiguration = this.guildConfigList.get(guildId) || {
+			allowAll: false,
+			announcementDelayHours: 5,
+			announcerMessage: null,
+			botname: "OmegaBot",
+			streamerChannelId: null,
+			streamerList: []
 		};
 
 		try {
 			const file = resolve(__dirname, "..", "infos", guildId + ".json");
 			try {
-				writeFileSync(file, JSON.stringify(GuildSettings, null, 2));
+				writeFileSync(file, JSON.stringify(GuildConfig, null, 2));
+				!msg ? null : msg.react("👍");
 			} catch (error) {
 				console.log(LOGTAG.ERROR, error);
+				!msg ? null : msg.react("👎");
 			}
 
 		} catch (error) {
 			console.log(LOGTAG.ERROR, error);
+			!msg ? null : msg.react("👎");
 		}
 	}
 
@@ -71,23 +88,25 @@ export class OmegaBot extends WorkerProcess {
 	 */
 	protected loadGuildSettings(guildId: string) {
 		if (this.settingsLoaded.has(guildId) && this.settingsLoaded.get(guildId)) return;
-		let GuildSettings = null;
+		let GuildConfig: GuildConfiguration = {
+			allowAll: false,
+			announcementDelayHours: 5,
+			announcerMessage: null,
+			botname: "OmegaBot",
+			streamerChannelId: null,
+			streamerList: []
+		};
 		try {
 			const file = resolve(__dirname, "..", "infos", guildId + ".json");
 			try {
 				const dataRaw = readFileSync(file);
-				GuildSettings = JSON.parse(dataRaw.toString());
-				this.streamerAllowAll.set(guildId, GuildSettings.allowAll);
-				this.streamerAllowed.set(guildId, GuildSettings.streamerList);
-				this.streamerChannel.set(guildId, GuildSettings.streamerChannelId);
+				GuildConfig = JSON.parse(dataRaw.toString());
 				!cfg.log.info ? null : console.log(LOGTAG.INFO, "[OmegaBot:loadGuildSettings]", `Guild <${guildId}> settings found and loaded!`);
 			} catch (error) {
 				!cfg.log.info ? null : console.log(LOGTAG.INFO, "[OmegaBot:loadGuildSettings]", `Guild <${guildId}> not found set all to default`);
-				this.streamerAllowAll.set(guildId, false);
-				this.streamerAllowed.set(guildId, []);
-				this.streamerChannel.set(guildId, null);
 			} finally {
 				this.settingsLoaded.set(guildId, true);
+				this.guildConfigList.set(guildId, GuildConfig);
 			}
 
 		} catch (error) {
@@ -112,28 +131,30 @@ export class OmegaBot extends WorkerProcess {
 				!cfg.log.info ? null : console.log(LOGTAG.INFO, "[OmegaBot:setupDiscordBot]", `Logged in as ${this.DiscordBot.user.tag}!`);
 
 				this.DiscordBot.guilds.forEach((G, key) => {
+					this.loadGuildSettings(G.id);
+					const { botname } = this.guildConfigList.get(G.id);
+					if (botname && G.me.hasPermission(Permissions.FLAGS.CHANGE_NICKNAME)) G.me.setNickname(botname);
 					!cfg.log.info ? null : console.log(LOGTAG.INFO, "[OmegaBot:setupDiscordBot]", `I'am member of ${G.name} with ${G.memberCount} members`);
 					if (!this.streamerChecks.has(key)) {
 						this.streamerChecks.set(key, setTimeout(() => {
 							const Guild: Guild = G;
-							this.loadGuildSettings(Guild.id);
 
 							// if (!this.streamerChannel.has(G.id)) {
 							// 	this.streamerChannel.set(G.id, null);
 							// }
-							const streamerChannelId = this.streamerChannel.get(G.id);
+							const { streamerChannelId, allowAll, streamerList, announcementDelayHours, announcerMessage } = this.guildConfigList.get(G.id);
 
 							if (!streamerChannelId) return;
 
 							// if (!this.streamerAllowAll.has(G.id)) {
 							// 	this.streamerAllowAll.set(G.id, false);
 							// }
-							const allowAll = this.streamerAllowAll.get(G.id);
+							// const allowAll = this.streamerAllowAll.get(G.id);
 
 							// if (!this.streamerAllowed.has(G.id)) {
 							// 	this.streamerAllowed.set(G.id, []);
 							// }
-							const allowedStreamer = this.streamerAllowed.get(G.id);
+							// const streamerList = this.streamerAllowed.get(G.id);
 
 							if (!this.announcementCache.has(G.id)) {
 								this.announcementCache.set(G.id, new Map<string, Game>());
@@ -144,17 +165,17 @@ export class OmegaBot extends WorkerProcess {
 							const aCache = this.announcementCache.get(G.id);
 							const aDateCache = this.announcementDateCache.get(G.id);
 							const blockTime = new Date();
-							blockTime.setHours(blockTime.getHours()-1);
+							blockTime.setHours(blockTime.getHours() - (announcementDelayHours || 5));
 
 							Guild.members.forEach((Member, key) => {
 								const Game = Member.presence.game;
 								const lastGame = aCache.get(Member.id);
 								const liveDate = aDateCache.get(Member.id);
-								if (Game && Game.streaming && (!lastGame || !lastGame.streaming) && (!liveDate || liveDate.getTime()<blockTime.getTime()) && (allowAll || allowedStreamer.includes(Member.id))) {
+								if (Game && Game.streaming && (!lastGame || !lastGame.streaming) && (!liveDate || liveDate.getTime() < blockTime.getTime()) && (allowAll || streamerList.includes(Member.id))) {
 									const txtCh: TextChannel = <TextChannel>Guild.channels.get(streamerChannelId);
 									try {
 										!txtCh ? null : txtCh.send(`@everyone Aufgepasst ihr Seelen! \`${Member.displayName}\` streamt gerade! \n\`${Game.name}\` - \`${Game.details}\` \n Siehe hier:${Game.url}`);
-										aDateCache.set(Member.id,new Date());
+										aDateCache.set(Member.id, new Date());
 									} catch (error) {
 										console.log(error);
 									}
@@ -181,162 +202,313 @@ export class OmegaBot extends WorkerProcess {
 					return;
 				}
 				const guildId: string = msg.guild.id;
+				const me = msg.guild.me;
 				const TC: TextChannel = msg.channel as TextChannel;
 
 				const author = msg.author;
 				const member = await msg.guild.fetchMember(author);
 				const isDeveloper = member.id == "385696536949948428";
 				const isAdmin = member.hasPermission("ADMINISTRATOR") || isDeveloper;
+				const GuildConfig = this.guildConfigList.get(guildId)
+				const { streamerChannelId, allowAll, streamerList, announcementDelayHours, announcerMessage } = GuildConfig;
 
+				const [command, args] = msg.content.split(" ", 2);
 
-				if (msg.content.startsWith("!remove")) {
-					if (isAdmin) {
-						const [command, target] = msg.content.split(" ");
-						const file = resolve(__dirname, "..", "infos", guildId, target.toLowerCase() + ".json");
-						try {
-							unlinkSync(file);
-							msg.react("👍");
-						} catch (error) {
-							msg.react("👎");
-						}
-					} else {
+				if (command.startsWith('!')) {
+					if (!isAdmin) {
 						msg.react("👎");
+						TC.send(`Du hast nicht die benötigten Rechte um \`${command}\` auszuführen, versuch es erst gar nicht! Probier doch mal \`?help\` `);
 						return;
 					}
-				} else if (msg.content.startsWith("!addStreamer")) {
-					if (isAdmin) {
-						const [command, streamer] = msg.content.split(" ");
-						const streamerList = this.streamerAllowed.get(guildId) || [];
-						msg.mentions.members.array().forEach((Member) => {
-							if (!streamerList.includes(Member.id)) {
-								streamerList.push(Member.id);
+					switch (command) {
+						case "!remove": {
+							const file = resolve(__dirname, "..", "infos", guildId, args.toLowerCase() + ".json");
+							try {
+								unlinkSync(file);
+								msg.react("👍");
+							} catch (error) {
+								msg.react("👎");
 							}
-						});
-						this.streamerAllowed.set(guildId, streamerList);
-						try {
-							this.saveGuildSettings(guildId);
-							msg.react("👍");
-						} catch (error) {
-							msg.react("👎");
-						}
-					} else {
-						msg.react("👎");
-						return;
-					}
-				} else if (msg.content.startsWith("!add")) {
-					if (isAdmin) {
-						const [command, target, ...text] = msg.content.split(" ");
-						if(target=="help") {
-							msg.react("👎");
-							return;
-						}
-						const file = resolve(__dirname, "..", "infos", guildId, target.toLowerCase() + ".json");
-						let data = null;
-						try {
-							const dataRaw = readFileSync(file);
-							data = JSON.parse(dataRaw.toString());
-							if (Array.isArray(data.data)) {
-								data.data.push(text.join(' '));
-							} else {
-								data.data = [data.data, text.join(' ')];
+						} break;
+						case "!add": {
+							const [target, text] = args.split(" ", 2);
+							if (target == "help") {
+								msg.react("👎");
+								return;
 							}
-						} catch (error) {
-							data = { data: text.join(' ') };
-						}
-						writeFileSync(file, JSON.stringify(data, null, 2));
-						msg.react("👍");
-						return;
-					} else {
-						msg.react("👎");
-						return;
-					}
-				} else if (msg.content.startsWith("!removeStreamer")) {
-					if (isAdmin) {
-						const [command, streamer] = msg.content.split(" ");
-						const streamerList = this.streamerAllowed.get(guildId) || [];
-						msg.mentions.members.array().forEach((Member) => {
-							if (streamerList.includes(Member.id)) {
-								streamerList.splice(streamerList.indexOf(Member.id), 1);
+							const file = resolve(__dirname, "..", "infos", guildId, target.toLowerCase() + ".json");
+							let data = null;
+							try {
+								const dataRaw = readFileSync(file);
+								data = JSON.parse(dataRaw.toString());
+								if (Array.isArray(data.data)) {
+									data.data.push(text);
+								} else {
+									data.data = [data.data, text];
+								}
+							} catch (error) {
+								data = { data: text };
 							}
-						});
-						this.streamerAllowed.set(guildId, streamerList);
-						try {
-							this.saveGuildSettings(guildId);
+							writeFileSync(file, JSON.stringify(data, null, 2));
 							msg.react("👍");
-						} catch (error) {
-							msg.react("👎");
-						}
-					} else {
-						msg.react("👎");
-						return;
-					}
-				} else if (msg.content.startsWith("!setStreamChannel")) {
-					if (isAdmin) {
-						this.streamerChannel.set(guildId, TC.id);
-						try {
-							this.saveGuildSettings(guildId);
-							msg.react("👍");
-						} catch (error) {
-							msg.react("👎");
-						}
-					} else {
-						msg.react("👎");
-						return;
+						} break;
+						case "!addStreamer": {
+							msg.mentions.members.array().forEach((Member) => {
+								if (!GuildConfig.streamerList.includes(Member.id)) {
+									GuildConfig.streamerList.push(Member.id);
+								}
+							});
+							this.saveGuildSettings(guildId, msg);
+						} break;
+						case "!removeStreamer": {
+							msg.mentions.members.array().forEach((Member) => {
+								if (GuildConfig.streamerList.includes(Member.id)) {
+									GuildConfig.streamerList.splice(GuildConfig.streamerList.indexOf(Member.id), 1);
+								}
+							});
+							this.saveGuildSettings(guildId, msg);
+						} break;
+						case "!setStreamChannel": {
+							GuildConfig.streamerChannelId = TC.id;
+							this.saveGuildSettings(guildId, msg);
+						} break;
+						case "!setAllowAll": {
+							const to = args == "true";
+							GuildConfig.allowAll = to;
+							this.saveGuildSettings(guildId, msg);
+						} break;
+						case "set": {
+							const [prop, options] = args.split(" ", 2);
+							switch (prop) {
+								case "allowAll": {
+									GuildConfig.allowAll = options == "true";
+									this.saveGuildSettings(guildId, msg);
+								} break;
+								case "name": {
+									GuildConfig.botname = options;
+									if (me.hasPermission(Permissions.FLAGS.CHANGE_NICKNAME)) {
+										me.setNickname(options);
+										TC.send(`Okay, dann heiße ich nun *${options}* für dich!`);
+										this.saveGuildSettings(guildId, msg);
+									} else {
+										msg.react("👎");
+										TC.send(`Tut mir leid, aber ich habe nicht das recht meinen Nicknamen selber zu ändern.`);
+									}
+								} break;
+								case "streamerChannel": {
+									GuildConfig.streamerChannelId = TC.id;
+									this.saveGuildSettings(guildId, msg);
+								} break;
+								case "announcementDelayHours": {
+									GuildConfig.announcementDelayHours = Number(options);
+									this.saveGuildSettings(guildId, msg);
+								} break;
+								default:
+									msg.react("👎");
+									break;
+							}
+						} break;
+						case "": { } break;
+						case "": { } break;
+						case "": { } break;
+						default:
+							TC.send(`Oh das tut mir leid, aber das Kommando \`${command}\` sagt mir leider gar nichts. Probier doch mal \`?help\` `);
+							break;
 					}
 
-				} else if (msg.content.startsWith("!setAllowAll")) {
-					if (isAdmin) {
-						const [command, value] = msg.content.split(" ");
-						const to = value == "true";
-						this.streamerAllowAll.set(guildId, to);
-						try {
-							this.saveGuildSettings(guildId);
-							msg.react("👍");
-						} catch (error) {
-							msg.react("👎");
-						}
-					} else {
-						msg.react("👎");
-						return;
-					}
-				} else if (msg.content.startsWith('?help')) {
-					const [command, value] = msg.content.split(" ");
-					return TC.send(`Oh, du hast die Kommandos vergessen? Hier Bitte:\n
+				} else {// if(command.startsWith('?'))
+					const what = msg.content.substr(1); // without ?
+					switch (what) {
+						case "help":
+							TC.send(`Oh, du hast die Kommandos vergessen? Hier Bitte:\n
 \`\`\`\n
 Kommandos für Administratoren:
 !add [was?] [text]          | Füge einen neuen text hinzu der per ?[was] wieder abgerufen werden kann, zum Beispiel Zitate oder Infos
 !remove [was?]              | Entferne alle Einträge zu [was] aus dem Speicher
 !setStreamChannel           | Der aktuelle Kanal wird zum Streamer Kanal, hier landen alle Ankündigungen
 !setAllowAll [true|false]   | Erlaube das ich jeden Streamer angekündigt darf [true] oder nicht [false]
-!addStreamer @name @name... | Füge ein oder mehrere Streamer hinzu die ich ankündigen darf!
-!removeStreamer @name       | Du kannst einen Streamer auch wieder entfernen, dann bleibe ich still
+!addStreamer @name ...      | Füge ein oder mehrere Streamer hinzu die ich ankündigen darf!
+!removeStreamer @name ...   | Du kannst einen Streamer auch wieder entfernen, dann bleibe ich still
 -------------------------------
 Kommandos für alle anderen:
 ?help                       | Wenn du diese Hilfe hier mal wieder brauchst, sag einfach bescheid :)
 ?[was?]                     | Ich werde dir zeigen was ich zu [was?] weiss, wenn ich nichts weiss, sag ichs dir auch ;)
 \n\`\`\``);
-				} else if (msg.content.startsWith('?')) {
-					const name = msg.content.substr(1); // without ?
-					const datadir = resolve(__dirname, "..", "infos", guildId);
-					try {
-						mkdirSync(datadir);
-					} catch (error) {
-						console.log(LOGTAG.ERROR, error);
-					}
-					const file = resolve(datadir, name.toLowerCase() + ".json");
-					try {
-						const dataRaw = readFileSync(file);
-						const data = JSON.parse(dataRaw.toString());
-						if (Array.isArray(data.data)) {
-							shuffle(data.data);
-							return TC.send(`Oha, zu ${name} fällt mir zum Beispiel das hier ein: \n${data.data[0]}`);
-						} else {
-							return TC.send(`Zu ${name} kann ich dir nur so viel sagen: \n${data.data}`);
-						}
-					} catch (error) {
-						return TC.send("Darüber weiss ich überhaupt gar nichts!");
+							break;
+
+						default: {
+
+							const datadir = resolve(__dirname, "..", "infos", guildId);
+							try {
+								mkdirSync(datadir);
+							} catch (error) {
+								console.log(LOGTAG.ERROR, error);
+							}
+							const file = resolve(datadir, what.toLowerCase() + ".json");
+							try {
+								const dataRaw = readFileSync(file);
+								const data = JSON.parse(dataRaw.toString());
+								if (Array.isArray(data.data)) {
+									shuffle(data.data);
+									TC.send(`Oha, zu ${what} fällt mir zum Beispiel das hier ein: \n${data.data[0]}`);
+								} else {
+									TC.send(`Zu ${what} kann ich dir nur so viel sagen: \n${data.data}`);
+								}
+							} catch (error) {
+								TC.send("Darüber weiss ich überhaupt gar nichts!");
+							}
+						} break;
 					}
 				}
+
+				// 				if (msg.content.startsWith("!remove")) {
+				// 					if (isAdmin) {
+				// 						const [command, target] = msg.content.split(" ");
+				// 						const file = resolve(__dirname, "..", "infos", guildId, target.toLowerCase() + ".json");
+				// 						try {
+				// 							unlinkSync(file);
+				// 							msg.react("👍");
+				// 						} catch (error) {
+				// 							msg.react("👎");
+				// 						}
+				// 					} else {
+				// 						msg.react("👎");
+				// 						return;
+				// 					}
+				// 				} else if (msg.content.startsWith("!addStreamer")) {
+				// 					if (isAdmin) {
+				// 						const [command, streamer] = msg.content.split(" ");
+				// 						// const streamerList = this.streamerAllowed.get(guildId) || [];
+				// 						msg.mentions.members.array().forEach((Member) => {
+				// 							if (!GuildConfig.streamerList.includes(Member.id)) {
+				// 								GuildConfig.streamerList.push(Member.id);
+				// 							}
+				// 						});
+				// 						// this.streamerAllowed.set(guildId, streamerList);
+				// 						try {
+				// 							this.saveGuildSettings(guildId);
+				// 							msg.react("👍");
+				// 						} catch (error) {
+				// 							msg.react("👎");
+				// 						}
+				// 					} else {
+				// 						msg.react("👎");
+				// 						return;
+				// 					}
+				// 				} else if (msg.content.startsWith("!add")) {
+				// 					if (isAdmin) {
+				// 						const [command, target, ...text] = msg.content.split(" ");
+				// 						if (target == "help") {
+				// 							msg.react("👎");
+				// 							return;
+				// 						}
+				// 						const file = resolve(__dirname, "..", "infos", guildId, target.toLowerCase() + ".json");
+				// 						let data = null;
+				// 						try {
+				// 							const dataRaw = readFileSync(file);
+				// 							data = JSON.parse(dataRaw.toString());
+				// 							if (Array.isArray(data.data)) {
+				// 								data.data.push(text.join(' '));
+				// 							} else {
+				// 								data.data = [data.data, text.join(' ')];
+				// 							}
+				// 						} catch (error) {
+				// 							data = { data: text.join(' ') };
+				// 						}
+				// 						writeFileSync(file, JSON.stringify(data, null, 2));
+				// 						msg.react("👍");
+				// 						return;
+				// 					} else {
+				// 						msg.react("👎");
+				// 						return;
+				// 					}
+				// 				} else if (msg.content.startsWith("!removeStreamer")) {
+				// 					if (isAdmin) {
+				// 						const [command, streamer] = msg.content.split(" ");
+				// 						// const streamerList = this.streamerAllowed.get(guildId) || [];
+				// 						msg.mentions.members.array().forEach((Member) => {
+				// 							if (GuildConfig.streamerList.includes(Member.id)) {
+				// 								GuildConfig.streamerList.splice(GuildConfig.streamerList.indexOf(Member.id), 1);
+				// 							}
+				// 						});
+				// 						// this.streamerAllowed.set(guildId, streamerList);
+				// 						try {
+				// 							this.saveGuildSettings(guildId);
+				// 							msg.react("👍");
+				// 						} catch (error) {
+				// 							msg.react("👎");
+				// 						}
+				// 					} else {
+				// 						msg.react("👎");
+				// 						return;
+				// 					}
+				// 				} else if (msg.content.startsWith("!setStreamChannel")) {
+				// 					if (isAdmin) {
+				// 						GuildConfig.streamerChannelId = TC.id;
+				// 						try {
+				// 							this.saveGuildSettings(guildId);
+				// 							msg.react("👍");
+				// 						} catch (error) {
+				// 							msg.react("👎");
+				// 						}
+				// 					} else {
+				// 						msg.react("👎");
+				// 						return;
+				// 					}
+
+				// 				} else if (msg.content.startsWith("!setAllowAll")) {
+				// 					if (isAdmin) {
+				// 						const [command, value] = msg.content.split(" ");
+				// 						const to = value == "true";
+				// 						GuildConfig.allowAll = to;
+				// 						try {
+				// 							this.saveGuildSettings(guildId);
+				// 							msg.react("👍");
+				// 						} catch (error) {
+				// 							msg.react("👎");
+				// 						}
+				// 					} else {
+				// 						msg.react("👎");
+				// 						return;
+				// 					}
+				// 				} else if (msg.content.startsWith('?help')) {
+				// 					const [command, value] = msg.content.split(" ");
+				// 					return TC.send(`Oh, du hast die Kommandos vergessen? Hier Bitte:\n
+				// \`\`\`\n
+				// Kommandos für Administratoren:
+				// !add [was?] [text]          | Füge einen neuen text hinzu der per ?[was] wieder abgerufen werden kann, zum Beispiel Zitate oder Infos
+				// !remove [was?]              | Entferne alle Einträge zu [was] aus dem Speicher
+				// !setStreamChannel           | Der aktuelle Kanal wird zum Streamer Kanal, hier landen alle Ankündigungen
+				// !setAllowAll [true|false]   | Erlaube das ich jeden Streamer angekündigt darf [true] oder nicht [false]
+				// !addStreamer @name ...      | Füge ein oder mehrere Streamer hinzu die ich ankündigen darf!
+				// !removeStreamer @name ...   | Du kannst einen Streamer auch wieder entfernen, dann bleibe ich still
+				// -------------------------------
+				// Kommandos für alle anderen:
+				// ?help                       | Wenn du diese Hilfe hier mal wieder brauchst, sag einfach bescheid :)
+				// ?[was?]                     | Ich werde dir zeigen was ich zu [was?] weiss, wenn ich nichts weiss, sag ichs dir auch ;)
+				// \n\`\`\``);
+				// 				} else if (msg.content.startsWith('?')) {
+				// 					const name = msg.content.substr(1); // without ?
+				// 					const datadir = resolve(__dirname, "..", "infos", guildId);
+				// 					try {
+				// 						mkdirSync(datadir);
+				// 					} catch (error) {
+				// 						console.log(LOGTAG.ERROR, error);
+				// 					}
+				// 					const file = resolve(datadir, name.toLowerCase() + ".json");
+				// 					try {
+				// 						const dataRaw = readFileSync(file);
+				// 						const data = JSON.parse(dataRaw.toString());
+				// 						if (Array.isArray(data.data)) {
+				// 							shuffle(data.data);
+				// 							TC.send(`Oha, zu ${name} fällt mir zum Beispiel das hier ein: \n${data.data[0]}`);
+				// 						} else {
+				// 							TC.send(`Zu ${name} kann ich dir nur so viel sagen: \n${data.data}`);
+				// 						}
+				// 					} catch (error) {
+				// 						TC.send("Darüber weiss ich überhaupt gar nichts!");
+				// 					}
+				// 				}
 
 
 
@@ -366,7 +538,7 @@ Kommandos für alle anderen:
 				// 	// 	WHO.file = new Attachment(buffer, "screenshot.jpg");
 				// 	// }
 
-					
+
 				// }).catch(e => {
 				// 	console.log(LOGTAG.ERROR, e);
 				// });